Mv: reorganize js into src/
