+++ index.html
@@ -40,6 +40,6 @@
 		<script src="https://cdnjs.cloudflare.com/ajax/libs/fuse.js/3.2.1/fuse.min.js"></script>
 		<!-- User created scripts -->
 		<script src="dist/database.js"></script>  <!-- Load the localDB -->
-		<script type="module" src="main.js"></script>
+		<script type="module" src="src/js/main.js"></script>
 	</body>
 </html>
