--- src/html/tmpl_bot.html
@@ -1,82 +1,0 @@
-				<footer class="footer row">
-					<div class="twelve columns" style="height: 10px;"></div>
-					<div class="twelve columns" style="text-align: center;">Navigation based on the <a href="https://codyhouse.co/gem/3d-bold-navigation/" title="CodyHouse">3D Bold Navigation</a> theme from CodyHouse and styled with <a href="http://getskeleton.com">Skeleton.css</a></div>
-					<div class="twelve columns" style="text-align: center;">Icons made by <a href="http://www.freepik.com" title="Freepik">Freepik</a>, <a href="https://www.flaticon.com/authors/coucou" title="Coucou">Coucou</a>, <a href="https://www.flaticon.com/authors/baianat" title="Baianat">Baianat</a>, <a href="https://www.flaticon.com/authors/appzgear" title="Appzgear">Appzgear</a>, and <a href="https://www.flaticon.com/authors/roundicons-freebies" title="Roundicons Freebies">Roundicons Freebies</a> from <a href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a> are licensed by <a href="http://creativecommons.org/licenses/by/3.0/" title="Creative Commons BY 3.0" target="_blank">CC 3.0 BY</a></div>
-					<div class="twelve columns" style="text-align: center;">Great Recipe Sources: <a href="https://cooking.nytimes.com/">New York Times</a></div>
-					<div class="twelve columns" style="height: 40px;"></div>
-				</footer>
-			</div> <!-- /container -->
-		</section> <!-- .cd-section -->
-	</main>
-
-	<nav class="cd-nav-container" id="cd-nav">
-		<header>
-			<h3>Navigation</h3>
-			<a href="#0" class="cd-close-nav">Close</a>
-		</header>
-
-		<ul class="cd-nav">
-			<li class="cd-unstyled-nav-item cd-selected" data-menu="top">
-				<a href="#toc">
-					<span>
-						<img class="nc-icon outline" src="src/icons/up-arrow.png" style="height: 64px; width: 64px;">
-					</span>
-					<em>table of contents</em>
-				</a>
-			</li>
-
-			<li class="cd-unstyled-nav-item" data-menu="brkfst">
-				<a href="#brkfst">
-					<span>
-						<img class="nc-icon outline" src="src/icons/croisant.png" style="height: 64px; width: 64px;">
-						<!-- <img class="nc-icon outline" src="src/icons/donut.png" style="height: 64px; width: 64px;"> -->
-					</span>
-					<em>brkfst</em>
-				</a>
-			</li>
-
-			<li class="cd-unstyled-nav-item" data-menu="dssrt">
-				<a href="#dssrt">
-					<span>
-						<img class="nc-icon outline" src="src/icons/cupcake.png" style="height: 64px; width: 64px;">
-					</span>
-					<em>dssrt</em>
-				</a>
-			</li>
-
-			<li class="cd-unstyled-nav-item" data-menu="grains">
-				<a href="#grains">
-					<span>
-						<img class="nc-icon outline" src="src/icons/rice-bowl.png" style="height: 64px; width: 64px;">
-					</span>
-					<em>grains</em>
-				</a>
-			</li>
-
-			<li class="cd-unstyled-nav-item" data-menu="dinner">
-				<a href="#dinner">
-					<span>
-						<img class="nc-icon outline" src="src/icons/bell-covering-hot-dish.png" style="height: 64px; width: 64px;">
-					</span>
-					<em>dinner</em>
-				</a>
-			</li>
-
-			<li class="cd-unstyled-nav-item" data-menu="sides">
-				<a href="#sides">
-					<span>
-						<img class="nc-icon outline" src="src/icons/dining-room-furniture-of-a-table-with-chairs.png" style="height: 64px; width: 64px;">
-					</span>
-					<em>sides</em>
-				</a>
-			</li>
-		</ul> <!-- .cd-3d-nav -->
-	</nav>
-
-	<div class="cd-overlay"><!-- shadow layer visible when navigation is visible --></div>
-
-<script src="src/js/jquery-2.1.1.js"></script>
-<script src="src/js/main.js"></script> <!-- Resource jQuery -->
-
-</body>
-</html>
