--- src/html/tmpl_top.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-		<title>kyleking/recipes</title>
-		<meta name="author" content="Kyle King">
-		<meta name="description" content="Basic Recipes">
-		<link rel="icon" href="src/favicon.ico">
-		<!-- Styles -->
-		<link href="src/styles/normalize.css" rel="stylesheet">
-		<link href="src/styles/skeleton.css" rel="stylesheet">
-		<link href="src/styles/checkbox.css" rel="stylesheet">
-		<link href="src/styles/codyhouse.css" rel="stylesheet">
-		<link href="src/styles/index.css" rel="stylesheet">
-
-		<!-- <link href='http://fonts.googleapis.com/css?family=Lora:400,700|Open+Sans:400,300,700' rel='stylesheet' type='text/css'> -->
-		<script src="src/js/modernizr.js"></script> <!-- Modernizr -->
-	</head>
-	<body>
-
-		<a href="#cd-nav" class="cd-nav-trigger">
-			Menu<span><!-- used to create the menu icon --></span>
-		</a> <!-- .cd-nav-trigger -->
-
-		<main>
-			<section class="cd-section index cd-selected">
-				<div class="container recipes cd-content">
-					<span id="top"></span>
