+++ src/js/Highlighter.js
